feat: a skeleton code to overly a text on an image
when predicting an image, I want to visualize the result by
overlaying a text on the position of the detected number.
In order to that, I need to have a piece of code to draw a
text, overlaying an image.

This is the piece of code implementing the feature I described.
It is just a skeleton version, no coding style, no comments,
needs futher modification to make it more readable